pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vector-os-nano"
version = "0.1.0"
description = "Natural language robot arm control SDK"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [{name = "Vector Robotics"}]
keywords = ["robotics", "ros2", "robot-arm", "llm", "natural-language", "so-101"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
]

dependencies = [
    "numpy>=1.24",
    "pyserial>=3.5",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "openai>=1.0",
]

[project.optional-dependencies]
# Perception: Camera + VLM support
perception = [
    "pyrealsense2>=2.50",
    "open3d>=0.17",
    "opencv-python>=4.8",
]

gpu = [
    "vector-os-nano[perception]",
]

# IK: Inverse kinematics solver
# pin (Pinocchio) >=3.9.0 for SO-101 URDF compatibility
ik = [
    "pin>=3.9.0",
]


# Simulation: MuJoCo (recommended) + PyBullet (legacy)
sim = [
    "mujoco>=3.0",
    "pybullet>=3.2",
]

# ROS2 integration (packages installed via apt, not pip)
ros2 = [
    # ros-humble-rclpy
    # ros-humble-launch-ros
    # ros-humble-sensor-msgs
]

mcp = [
    "mcp>=1.0",
]

# Go2 quadruped locomotion: Pinocchio kinematics + CasADi for MPC
go2 = ["casadi>=3.6", "pin>=3.0"]

all = [
    "vector-os-nano[perception,ik,sim,mcp]",
]

# Development tools
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "lark",
]

[project.scripts]
vector-os-mcp = "vector_os_nano.mcp.server:main_sync"
vector-cli = "vector_os_nano.vcli.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["vector_os_nano"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-p no:launch_ros -p no:launch_testing -p no:ament_copyright -p no:ament_flake8 -p no:ament_pep257 -p no:ament_xmllint -p no:ament_lint"
markers = [
    "level0: Level 0 — MuJoCo physics validity (no MuJoCoGo2 wrapper required)",
    "level1: Level 1 — PD standing stability",
    "level2: Level 2 — Open-loop locomotion / walking",
    "level3: Level 3 — Velocity tracking via set_velocity()",
    "level4: Level 4 — Navigation to waypoint (P-controller placeholder)",
    "ros2: requires rclpy installed and a live DDS context",
]
